+++ src/Drawing.SkiaSharp/Utilities/DrawUtility.cs
@@ -32,8 +32,8 @@ public static class DrawUtility
     {
         var images = imageLayers.ToList();
         var size = new SKSize(
-            (int)images.Max(x => x.Options?.Size?.Width ?? (x.Source.Size?.Width ?? 0)),
-            (int)images.Max(x => x.Options?.Size?.Height ?? (x.Source.Size?.Height ?? 0))
+            images.Max(x => x.Options?.Size?.Width ?? (x.Source.Size?.Width ?? 0)),
+            images.Max(x => x.Options?.Size?.Height ?? (x.Source.Size?.Height ?? 0))
         );
         return SkiaUtility.Create(size, ImageDefaults.BackgroundColor.ToSkiaColor());
     }
